# compiler, ci: remove acyclic and run leak tests