--- a/compiler/ast/ast_parsed_types.nim
+++ b/compiler/ast/ast_parsed_types.nim
@@ -226,7 +226,7 @@ type
     base*: NumericalBase
     literal*: string
 
-  ParsedNodeData*{.final, acyclic.} = object
+  ParsedNodeData*{.final.} = object
     # TODO: replace token fields with indexing into a token sequence, this
     #       should also address line info tracking.
     comment*: string       # TODO: replace with an index into a token stream