--- a/compiler/backend/cgir.nim
+++ b/compiler/backend/cgir.nim
@@ -169,7 +169,7 @@ type
   LocalId* = distinct uint32
     ## Identifies a local within a procedure.
 
-  CgNode* = ref object
+  CgNode* {.acyclic.} = ref object
     ## A node in the tree structure representing code during the code
     ## generation stage. The "CG" prefix is short for "code generation".
     info*: TLineInfo