--- a/compiler/utils/btrees.nim
+++ b/compiler/utils/btrees.nim
@@ -16,7 +16,7 @@ const
   Mhalf = M div 2
 
 type
-  Node[Key, Val] {.acyclic.} = ref object
+  Node[Key, Val]  = ref object
     entries: int
     keys: array[M, Key]
     case isInternal: bool