--- a/compiler/utils/btrees.nim
+++ b/compiler/utils/btrees.nim
@@ -16,6 +16,7 @@ const
   Mhalf = M div 2
 
 type
+  # TODO: This type should be marked acyclic
   Node[Key, Val]  = ref object
     entries: int
     keys: array[M, Key]