@alaviss 2024-01-18
As an aside, here are the numbers with ORC cycle collector enabled for the compiler.

```
        Time             Ratio
Before  9.155s  ± 0.044s  1.00
After   9.599s  ± 0.039s  1.05
ORC on  11.417s ± 0.150s  1.25
```