@alaviss 2024-01-18
> As for removing `.acyclic` from all types, I strongly disagree. Most reference types used in the compiler are indeed acyclic, even though it doesn't look like it from their type definitions.

While I agree that these annotations might return, I think it's best that we have this PR bring us back to a "correct" baseline, then reintroduce the annotations in later PRs.

This makes sure that we have a documented trail of progress as well as information about why those types are acyclic documented within the compiler commit history.

I'm a strong believer of make it work, make it correct then make it fast, so I'd prefer each steps to be documented.

As an aside, it seems like this change brought out some allocator bugs which caused tcompilerapi to crash, so we will have to look into that first.