@zerbina 2024-01-18
> I'm a strong believer of make it work, make it correct then make it fast, so I'd prefer each steps to be documented.

But removing all `.acyclic` annotations is a step back, in that it *hides* potential issues. If, for example, `ParsedDataNode` forms a cycle, then that's flat out wrong, but it's no longer caught now. 

Also, while it's true that the other types have no dedicated commit that I know of, `CgNode` *does* have a dedicated PR that marked it as `.acyclic`.