@zerbina 2024-01-18
> As an aside, it seems like this change brought out some allocator bugs which caused tcompilerapi to crash, so we will have to look into that first.

I've looked into it, and it's not a bug with the allocator, rather a fundamental issue with how the cycle collector handles destruction of cells. It's a variation of the issue with `.acyclic` I mentioned -- I'll create the issue today.