@alaviss 2024-01-18
> But removing all `.acyclic` annotations is a step back, in that it _hides_ potential issues. If, for example, `ParsedDataNode` forms a cycle, then that's flat out wrong, but it's no longer caught now.

Right now, `{.acyclic.}` is an optimization, not an indicator for correctness. While this PR introduces a process to verify these using sanitizers, I would not consider it reliable on any platforms other than Linux. As far as we know of right now, any of these could have not met the assertion, as they were not tested until now.

Which is why I think we should move ahead slow and steady, reintroduce these annotations with reasoning in code in future PR to contain those changes with a vote of confidence from CI.

Ideally we add a mode in ORC to verify that "acyclic" types don't exhibit cycles so local users can test it easily, but that's work for the future.

Furthermore, as you noted, there are issues with how the runtime handles acyclic as well, so it doesn't inspire confidence that acyclic is safe to apply without proper reasoning and testing.