@zerbina 2024-01-18
> Right now, `{.acyclic.}` is an optimization, not an indicator for correctness

I somewhat disagree. I see it as an assertion made by the programmer to the compiler that the type isn't really cyclic even though it looks like it. While the compiler has to trust this assertion, although it could - as you said - verify it at runtime, the leak tests act as a strong verification of assumption now. 

> Furthermore, as you noted, there are issues with how the runtime handles acyclic as well

That's not what I said. There are issue with how the runtime handles freeing of cells, but that's not directly related to `.acyclic`.

---

As a compromise, what do you think about the following:
* the `.acylic` pragma is not removed for `CgNode` (refer to https://github.com/nim-works/nimskull/pull/992)
* I fix the cycle collector issue
* I, after this PR is merged, make a PR that reinstates the `.acyclic` pragma in all cases I mentioned