@saem 2024-01-18
> Furthermore, as you noted, there are issues with how the runtime handles acyclic as well, so it doesn't inspire confidence that acyclic is safe to apply without proper reasoning and testing.

Besides `TNode` and friends that are not in fact `acyclic`, the others ones are in terms of their intention. So we can apply `acyclic` based on the behaviour of the implementation or based on intention. I believe @zerbina's perspective is that we should do it by intention and yours is by behaviour.

Going by behaviour has the plus side of having the correct net effect, while taking into account the current implementation's short comings. The downside is that the implementation is a moving target and so is the code that's using it, then we've got two moving targets, plus some tests which may or may not provide a sufficient spec (with failures marked as known issues). Also, since this stuff is lots of threading the needle, one has to actively remember where all the dragons lie, which isn't easy if there isn't directly focused effort on the collector & destructors.

Going by the intention has the plus side of testing the interface as it should be, and there is only one moving target the implementation, while the compiler code and tests together act as a spec (higher likelihood of it being more complete). The downside of course is that we might have some serious issues with the implementation so we _might_ have to compromise on intention correct code, but that's not guaranteed.

I personally favour the intention approach, because that exercises things with and without `acyclic`, and we can note leaving those types as such in this PR. Not to mention, it's easier to remember to take something away, than it is to re-add something. Finally, since working on arc/orc is more indirect work right now based on as the need arises, rather than dedicated effort, the behaviouar based approach is more onerous in terms of having to remember all sorts of edge cases whenever usage gets revisited.