@zerbina 2024-01-18
> > As a compromise, what do you think about the following
> 
> Isn't this how I was advocating everything to be done? :P

It's not how I understood your suggestion. You didn't mention `CgNode` and you also said "slow and steady", which I consider a single PR to be the opposite of.