@zerbina 2024-01-20
There was a bit of back and forth on this topic now. What's the agreed-on plan now?

If the correct `.acyclic` are removed, then the places I mentioned should at least be marked with a comment to note that `.acyclic` needs to be restored.