@alaviss 2024-01-20
> If the correct `.acyclic` are removed, then the places I mentioned should at least be marked with a comment to note that `.acyclic` needs to be restored.

sure, I can do that