@saem 2024-01-20
> I still think that these annotations should be restored separately from this PR. For two reasons:
> 
> * As mentioned before, this PR brings us back to a "correct" (read: leakless) state.
> * In the restore PR there would be more space to do things like adding an explainer on why is something acyclic, which I think should be done for all types marked as such.

If you can run with the above plan then I'm good with it, one less thing for zerbina or I to keep in mind.